+++ packages/guide/src/app/database/components/database-offerings/database-offerings.component.ts
@@ -14,6 +14,7 @@ export class DatabaseOfferingsComponent extends BaseDatabaseDetailPartComponent 
     ngOnInit(): void {
         if (!this.item) return;
         const recipes = this.database.getOfferings();
+        const tagBasedItems = this.database.getTagBasedItems();
 
         this.isRewardIn = recipes.map(altar => {
             const offerings = altar.offerings.filter(offering => offering.rewards.items.find(reward => reward.item.id === this.item?.id) || offering.rewards.recipes.find(reward => reward.item.id === this.item?.id));
@@ -22,7 +23,16 @@ export class DatabaseOfferingsComponent extends BaseDatabaseDetailPartComponent 
         }).filter(nonNullable);
 
         this.requiredAsOffering = recipes.map(altar => {
-            const offerings = altar.offerings.filter(offering => offering.requiredItems.find(reward => reward.item.id === this.item?.id));
+            const offerings = altar.offerings.filter(offering => offering.requiredItems.find(reward => {
+                if ('id' in reward.item) {
+                    return reward.item.id === this.item?.id;
+                } else {
+                    const key = reward.item.key;
+                    const items = this.database.getTagBasedItems().find(t => key === t.key)?.items;
+
+                    return items?.find(t => t.id === this.item?.id)
+                }
+            }));
             if (!offerings.length) return null;
             return {...altar, offerings}
         }).filter(nonNullable);