+++ packages/guide/src/app/my-coral-guide/components/offerings-checklist/offerings-checklist.component.ts
@@ -1,22 +1,23 @@
 import { Component, inject } from '@angular/core';
 import { BaseTabbedSelectableContainerComponent } from "../../../shared/components/base-tabbed-selectable-container/base-tabbed-selectable-container.component";
-import { MinimalItem, Offering, OfferingAltar, Offerings } from "@ci/data-types";
+import { MinimalItem, MinimalTagBasedItem, Offering, OfferingAltar, Offerings } from "@ci/data-types";
 import { Observable, tap } from "rxjs";
 import { ToDoCategory } from "../../../core/enums/todo-category.enum";
 import { OfferingChecklistService } from "../../../core/services/checklists/offering-checklist.service";
 import { FormControl, FormRecord } from "@angular/forms";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
+import { entityKey } from "@ci/util";
 
 @Component({
     selector: 'app-offerings-checklist',
     templateUrl: './offerings-checklist.component.html',
 })
-export class OfferingsChecklistComponent extends BaseTabbedSelectableContainerComponent<MinimalItem> {
+export class OfferingsChecklistComponent extends BaseTabbedSelectableContainerComponent<MinimalItem | MinimalTagBasedItem> {
     checklistService = inject(OfferingChecklistService);
     checklistForm: FormRecord<FormControl<boolean>> = new FormRecord<FormControl<boolean>>({})
     protected activeOffering?: Offerings;
     protected offerings$: Observable<OfferingAltar[]>;
-    protected entryForToDo?: Offering | MinimalItem;
+    protected entryForToDo?: Offering | MinimalItem | MinimalTagBasedItem;
     protected toDoCategory = ToDoCategory;
 
     constructor() {
@@ -42,7 +43,8 @@ export class OfferingsChecklistComponent extends BaseTabbedSelectableContainerCo
                         const keys = Object.keys(checklist);
                         checklist.offerings.forEach(key => {
                             key.requiredItems.forEach(item => {
-                                this.checklistForm.addControl(item.item.id, new FormControl<boolean>(this.checklistService.isChecked(item.item.id), {nonNullable: true}), {emitEvent: false})
+                                const key = entityKey(item.item);
+                                this.checklistForm.addControl(key, new FormControl<boolean>(this.checklistService.isChecked(key), {nonNullable: true}), {emitEvent: false})
                             });
                         })
                     })
@@ -53,13 +55,13 @@ export class OfferingsChecklistComponent extends BaseTabbedSelectableContainerCo
 
     }
 
-    override registerToToDo(entry: MinimalItem | Offering) {
+    override registerToToDo(entry: MinimalItem | Offering | MinimalTagBasedItem) {
         if ('item' in entry) {
             this._todo.add(ToDoCategory.OFFERINGS, entry)
         }
     }
 
-    override showDetails(selectedEntry?: Offering | MinimalItem) {
+    override showDetails(selectedEntry?: Offering | MinimalItem | MinimalTagBasedItem) {
         this.entryForToDo = selectedEntry;
 
         if (selectedEntry && 'amount' in selectedEntry) {