+++ packages/guide/src/app/shared/pipes/to-item-list-entries.pipe.spec.ts
@@ -1,0 +1,8 @@
+import { ToItemListEntriesPipe } from './to-item-list-entries.pipe';
+
+describe('ToItemListEntriesPipe', () => {
+    it('create an instance', () => {
+        const pipe = new ToItemListEntriesPipe();
+        expect(pipe).toBeTruthy();
+    });
+});