+++ packages/guide/src/app/shared/pipes/to-item-list-entries.pipe.ts
@@ -1,0 +1,16 @@
+import { Pipe, PipeTransform } from '@angular/core';
+import { GenericEntry, Item, ItemEntry, MinimalItem } from "@ci/data-types";
+
+@Pipe({
+    name: 'toItemListEntries',
+})
+export class ToItemListEntriesPipe implements PipeTransform {
+    transform(value: (MinimalItem | Item)[]): (ItemEntry | GenericEntry)[] {
+        return value.map(i => {
+            return {
+                item: i,
+                amount: 1
+            }
+        });
+    }
+}