+++ packages/guide/src/app/shared/shared.module.ts
@@ -58,6 +58,10 @@ import { RemoveTagsPipe } from './pipes/remove-tags.pipe';
 import { ChancePerItemListComponent } from './components/chance-per-item-list/chance-per-item-list.component';
 import { ChancePerItemTableListComponent } from './components/chance-per-item-table-list/chance-per-item-table-list.component';
 import { HasPipe } from './pipes/has.pipe';
+import { EntityKeyPipe } from './pipes/entity-key.pipe';
+import { IsMinimalTagBasedItemPipe } from './pipes/is-minimal-tag-based-item.pipe';
+import { IsTagBasedItemPipe } from './pipes/is-tag-based-item.pipe';
+import { ToItemListEntriesPipe } from './pipes/to-item-list-entries.pipe';
 
 @NgModule({
     declarations: [
@@ -109,6 +113,10 @@ import { HasPipe } from './pipes/has.pipe';
         ChancePerItemListComponent,
         ChancePerItemTableListComponent,
         HasPipe,
+        EntityKeyPipe,
+        IsMinimalTagBasedItemPipe,
+        IsTagBasedItemPipe,
+        ToItemListEntriesPipe,
     ],
     imports: [
         CommonModule,
@@ -169,6 +177,10 @@ import { HasPipe } from './pipes/has.pipe';
         RemoveTagsPipe,
         ChancePerItemListComponent,
         ChancePerItemTableListComponent,
+        HasPipe,
+        EntityKeyPipe,
+        IsMinimalTagBasedItemPipe,
+        IsTagBasedItemPipe,
     ],
 })
 export class SharedModule {}