+++ packages/pak-files-parser/src/interfaces/raw-data-interfaces/raw-offering.interface.ts
@@ -1,15 +1,14 @@
+import { SourceString } from "../../types/source-string.type";
+
 export interface RawOffering {
-    "offeringTitleText": {
-        "Namespace": string
-        "Key": string
-        "SourceString": string
-    },
+    "offeringTitleText": SourceString,
     "offeringImage": {
         "ObjectName": string
         "ObjectPath": string
     },
     "requiredItems": [
         {
+            "useGenericItem": boolean,
             "itemData": {
                 "data": {
                     "DataTable": {
@@ -20,6 +19,13 @@ export interface RawOffering {
                 },
                 "itemID": string
             },
+            "genericItem": {
+                "genericItem": {
+                    "DataTable": string | null
+                    "RowName": string;
+                },
+                "amount": number
+            }
             "itemQuantity": number
         },
     ],